Move unused server modules into utils/deprecated.
These files will not be used by the proxy server; move them so the old HTTP server can be deleted as a unit later.
