+++ python/cuopt_server/cuopt_server/utils/deprecated/__init__.py
@@ -1,0 +1,6 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+# Legacy-only cuOpt HTTP server modules. Permanent utils must not import
+# this package. Deleting the old server is: remove cuopt_service.py,
+# webserver.py, this package, and tests/deprecated/.
