+++ python/cuopt_server/cuopt_server/utils/deprecated/linear_programming/generate_metadata.py
