+++ python/cuopt_server/cuopt_server/utils/deprecated/routing/generate_metadata.py
